requirements.txt
# Arquivo: requirements.txt

# Framework Web
Flask
Flask-Cors

# Banco de Dados
pymongo[srv]

# Autenticação (para as próximas seções)
passlib[bcrypt]
python-jose[cryptography]
Flask-JWT-Extended

# Manipulação de Arquivos
python-docx
docxtpl
openpyxl
pandas
reportlab

# IA
google-generativeai

# Configuração
python-dotenv

# Criptografia
bcrypt
